requirements.txt
psycopg2
defusedxml==0.4.1
Django==1.10.2
django-allauth==0.28.0
django-rest-auth==0.8.1
djangorestframework==3.4.7
oauthlib==2.0.0
python3-openid==3.0.10
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
django-oauth-toolkit==0.10.0
ipdb==0.10.1
model-mommy==1.2.6
